+++ tests/exceptions/test_manager.py
@@ -1,0 +1,10 @@
+import unittest
+
+from mkart.exceptions.manager import ManagerException
+
+
+class TestManagerException(unittest.TestCase):
+
+    def test_should_be_instance_from_exception(self):
+        manager_exception = ManagerException()
+        self.assertTrue(isinstance(manager_exception, Exception))